+++ src/core/reader.h
@@ -1,16 +1,8 @@
-#ifndef _PROTOTYPE_READER_H_
-#define _PROTOTYPE_READER_H_
+#ifndef _MINIM_READER_H_
+#define _MINIM_READER_H_
 
-#include <stdbool.h>
-#include <stdint.h>
+#include "ast.h"
 
-struct SyntaxTable
-{
-    size_t idx, row, col;
-    uint8_t flags;
-    char eof;
-} typedef SyntaxTable;
 
-SyntaxNode *minim_read_str(FILE *file, char eof, char wait);
 
 #endif
